third_party/WebKit/LayoutTests/inspector/tracing/timeline-misc/timeline-event-causes-expected.txt

Test that causes are correctly generated for various types of events.


Running: testTimerInstall
PASS - record contained Timer InstalledPromise @ setTimeoutFunction.js:

Running: testRequestAnimationFrame
PASS - record contained Animation Frame RequestedPromise @ requestAnimationFrameFunction.js:

Running: testStyleRecalc
PASS - record contained First InvalidatedstyleRecalcFunction @ styleRecalcFunction.js:

Running: testLayout
PASS - record contained Layout ForcedlayoutFunction @ layoutFunction.js:
PASS - record contained First Layout InvalidationlayoutFunction @ layoutFunction.js:
